requirements.txt
scikit-learn==0.24.1
fire==0.4.0
matplotlib==3.4.1
tqdm==4.49.0
python-mnist==0.7
numpy==1.22.3
pandas
numba

